add token validation and user creation tests

+++ part4/bloglist/controllers/blogs.js
@@ -1,9 +1,19 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+const getTokenFrom = req => {
+  const auth = req.get('authorization')
+  if (auth && auth.toLowerCase().startsWith('bearer ')) {
+    return auth.substring(7)
+  }
+  return null
+}
 
 blogsRouter.get('/', async (req, res) => {
   const blogs = await Blog.find({})
+    .populate('user', { username: 1, name: 1, id: 1 })
 
   res.send(blogs.map(blog => blog.toJSON()))
 })
@@ -11,7 +21,13 @@ blogsRouter.get('/', async (req, res) => {
 blogsRouter.post('/', async (req, res) => {
   const { body } = req
 
-  const user = await User.findById(body.userId)
+  const token = getTokenFrom(req)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     title: body.title,
@@ -21,13 +37,28 @@ blogsRouter.post('/', async (req, res) => {
     user: user._id
   })
   const savedBlog = await blog.save()
-  user.notes = user.notes.concat(savedBlog._id)
+  user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
 
   res.status(201).json(savedBlog.toJSON())
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
+  const token = getTokenFrom(req)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(req.params.id)
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return res.status(401).json({ error: 'wrong token' })
+  }
+  user.blogs = user.blogs.filter(b => b._id.toString() !== req.params.id.toString())
+  await user.save()
+
   await Blog.findByIdAndRemove(req.params.id)
   res.status(204).end()
 })
